Consolidating Bytetag block, Responsive updates
