+++ blocks/bytetag-buy/bytetag-buy.css
@@ -1,0 +1,56 @@
+.bytetag-buy-container > .default-content-wrapper {
+    text-align: center;
+}
+.bytetag-buy-container > .default-content-wrapper a {
+    color: var(  --number-background-color);
+    text-decoration: underline;
+}
+.bytetag-buy-wrapper {
+    display: flex;
+    justify-content: center;
+    width: 100%;
+    @media (min-width: 1024px){
+        padding: 2rem 0;
+    }
+}
+.bytetag-buy {
+    display: flex !important;
+    flex-direction: row-reverse !important;
+    max-width: 100% !important;
+    gap: 150px;
+    text-align: left;
+    @media (max-width: 600px){
+        flex-direction: column-reverse!important;
+        width: 100%;
+        gap: 25px;
+    }
+}
+.bytetag-buy > div {
+    width: 100%;
+}
+.bytetag-buy > div h2 {
+    font-size: 3rem;
+}
+.bytetag-buy > div h3 {
+    border-top: 1px solid #D9D9D9;
+    margin-top: 1rem;
+    padding-top: 1rem;
+    font-size: 1.625rem;
+    line-height: 2.25rem;
+}
+.bytetag-buy > div p {
+    font-size: 18px;
+    line-height:28px;
+}
+.bytetag-buy > div h3:first-of-type {
+    border-top: 0;
+}
+.bytetag-buy > div > div > p > picture {
+    display: inline-block;
+    margin-right: 6px;
+}
+.bytetag-buy > div > div > p > picture + * {
+    font-style: normal;
+    font-size: 24px;
+    line-height: 34px;
+}
