--- blocks/bytetag-slide/bytetag-slide.css
@@ -1,31 +1,0 @@
-.bytetag-slide-wrapper {
-    display: flex;
-    justify-content: center;
-    width: 100%;
-}
-.bytetag-slide {
-    display: flex !important;
-    flex-direction: row !important;
-    max-width: 100% !important;
-    gap: 150px;
-}
-@media (max-width: 600px){
-    .bytetag-slide {
-        flex-direction: column !important;
-        width: 100%;
-    }
-}
-.bytetag-slide > div {
-    max-width: 550px;
-}
-.bytetag-slide > iframe {
-    overflow-y: visible;
-}
-.bytetag-slide > div h3 {
-    border-top: 1px solid #D9D9D9;
-    margin-top: 1rem;
-    padding-top: 1rem;
-}
-.bytetag-slide > div h3:first-of-type {
-    border-top: 0;
-}
