+++ blocks/columns/columns.css
@@ -57,23 +57,43 @@
   border-radius: 1rem;
   color: var(  --text-color);
   border: 1px solid #D0D0D6;
-  gap: 16px;
+  gap: 8px;
   padding-bottom: 24px;
 }
 .columns.cards > div > div > p:first-of-type > picture > img:first-of-type {
   width: 100%;
   object-fit: cover;
   border-collapse: separate;
+  margin-bottom: 10px;
 }
-.columns.cards > div > div > h3, 
-.columns.cards > div > div > h3 > em{
+.columns.cards > div > div > h3 {
+  font-size: 20px;
+  line-height: 28px;
+}
+.columns.cards > div  > div > h3::after {
+  display: inline-block;
+  content: '';
+  margin-left: .8rem;
+  margin-bottom: 4px;
+  height: 16px;
+  width: 16px;
+  background: var(--text-color);
+  -webkit-mask: url('../../icons/circle-check.svg') no-repeat center center;
+  mask: url('../../icons/circle-check.svg') no-repeat center center;
+  vertical-align: middle;
+} 
+.columns.cards > div > div > h3{
   color: var(--text-color);
 }
+.columns.cards > div > div > p > em {
+  font-style: normal;
+  color: #666A6C;
+  font-size: 14px;
+  line-height: 28px;
+}
 .columns.cards > div > div > p:not(:first-of-type){
-  /* text-align: center; */
-  font-size: var(   --body-font-size-default);
+  font-size: var(--body-font-size-default);
   margin-bottom: var(--body-font-size-default);
-  /* padding: 0 16px; */
 }
 .columns.cards > div > div > p:not(:first-of-type),
 .columns.cards > div > div > h3 {
@@ -82,17 +102,26 @@
 }
 .columns.cards > div > div > p:last-of-type{
   margin-bottom: 0;
+  font-size: 16px;
+  line-height: 24px;
 }
-
 .section.highlight > div > .columns.cards > div > div {
   align-items: center;
   padding: 32px;
   box-shadow: 0px 4px 16px #000;
   background-color: var(  --background-color);
 }
+.section.highlight > div > .columns.cards > div > div h3 {
+  font-size: 26px;
+  line-height: 36px;
+}
+.section.highlight > div > .columns.cards > div > div h3::after {
+  content: none;
+}
 .section.highlight > div > .columns.cards > div > div p{
   text-align: center;
-  /* object-fit: unset; */
+  font-size: 18px;
+  line-height: 28px;
 }
 .columns > div > .columns-img-col {
   order: 0;
