+++ templates/bytetag/bytetag.js
@@ -1,0 +1,22 @@
+export async function loadLazy(main) {
+    const hero = document.createElement('div');
+    const imgDiv = document.createElement('div');
+    const contentDiv = document.createElement('div');
+    const defaultContentWrapper = main.querySelector('.default-content-wrapper');
+    const heroContainer = main.querySelector('.section.hero-container');
+    hero.className = 'hero-wrapper';
+    imgDiv.className = 'img-div';
+    contentDiv.classList = 'text-div';
+    [...heroContainer.querySelectorAll('picture')].forEach((el) => {
+      imgDiv.append(el);
+    });
+    hero.append(imgDiv);
+    contentDiv.append(document.querySelector('h1'));
+    [...defaultContentWrapper.querySelectorAll('p')].forEach((el) => {
+      if (el.innerText.trim() !== '') {
+        contentDiv.append(el);
+      }
+    });
+    hero.append(contentDiv);
+    heroContainer.replaceWith(hero);
+  }
